chore: PR fixes
